requirements.txt
transformers>=4.45.0
requests>=2.32.3
tiktoken>=0.7.0
verovio>=4.3.1
accelerate>=0.34.2
pdf2image>=1.17.0
python-multipart>=0.0.10
pypandoc>=1.13
pandoc>=2.4
# latex2markdown>=0.2.1
# pypdf2>=3.0.1
torchvision>=0.19.1

# sudo apt-get install poppler-utils pandoc
# pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
